requirements.txt
torch==2.0.1
transformers==4.30.2
numpy==1.24.3
pandas==2.0.2
scipy==1.10.1
biopython==1.81
matplotlib==3.7.1
plotly==5.15.0
seaborn==0.12.2
scikit-learn==1.3.0
fastapi==0.100.0
uvicorn==0.23.2
pydantic==2.0.3
pyyaml==6.0
networkx==3.1
torchvision==0.15.2
rdkit-pypi==2022.9.5
prody==2.4.0